requirements.txt
anyio
bcrypt
click
dnspython
ecdsa
email-validator
fastapi
h11
idna
passlib
Pillow
pyasn1
pydantic
pydentic
python-jose
python-multipart
python-slugify
python-stdnum
rsa
six
sniffio
SQLAlchemy
starlette
text-unidecode
typing_extensions
uvicorn